+++ crates/astria-cli/src/commands/sequencer.rs
@@ -151,6 +151,8 @@ pub(crate) async fn get_block_height(args: &BlockHeightGetArgs) -> eyre::Result<
 /// * If the http client cannot be created
 /// * If the latest block height cannot be retrieved
 pub(crate) async fn send_transfer(args: &TransferArgs) -> eyre::Result<()> {
+    use proto::native::sequencer::v1alpha1::asset;
+
     // Build the signing_key
     let private_key_bytes: [u8; 32] = hex::decode(args.private_key.as_str())
         .wrap_err("failed to decode private key bytes from hex string")?
@@ -178,9 +180,9 @@ pub(crate) async fn send_transfer(args: &TransferArgs) -> eyre::Result<()> {
         actions: vec![Action::Transfer(TransferAction {
             to: to_address,
             amount: args.amount,
-            asset_id: proto::native::sequencer::asset::default_native_asset_id(),
+            asset_id: asset::default_native_asset_id(),
         })],
-        fee_asset_id: proto::native::sequencer::asset::default_native_asset_id(),
+        fee_asset_id: asset::default_native_asset_id(),
     }
     .into_signed(&sequencer_key);
     let res = sequencer_client
