+++ crates/astria-proto/src/native/sequencer/mod.rs
@@ -1,2 +1,1 @@
-pub mod asset;
 pub mod v1alpha1;
