+++ crates/astria-proto/src/native/sequencer/v1alpha1/asset.rs
